Moving on to beta2.

+++ src/timeout.h
@@ -10,19 +10,19 @@
 
 /* timeout control structure */
 typedef struct t_tm_ {
-    int total;          /* total number of miliseconds for operation */
-    int block;          /* maximum time for blocking calls */
-    int start;          /* time of start of operation */
+    double total;          /* total number of miliseconds for operation */
+    double block;          /* maximum time for blocking calls */
+    double start;          /* time of start of operation */
 } t_tm;
 typedef t_tm *p_tm;
 
 int tm_open(lua_State *L);
-void tm_init(p_tm tm, int block, int total);
-int tm_get(p_tm tm);
-int tm_getretry(p_tm tm);
+void tm_init(p_tm tm, double block, double total);
+double tm_get(p_tm tm);
+double tm_getretry(p_tm tm);
 p_tm tm_markstart(p_tm tm);
-int tm_getstart(p_tm tm);
-int tm_gettime(void);
+double tm_getstart(p_tm tm);
+double tm_gettime(void);
 int tm_meth_settimeout(lua_State *L, p_tm tm);
 
 #endif /* TM_H */
